requirements.txt
pandas==1.3.5
Pillow==8.4.0
pymongo==4.0.1
pyquery==1.4.3
selenium==4.1.0
XlsxWriter==3.0.2
